Remove some extraneous casts to ssize_t since it doesn't appear that MUSL pulls it in

+++ src/hash/duncecap.c
@@ -142,7 +142,7 @@ duncecap_init_size(duncecap_t *self, char size)
     duncecap_store_t *store;
     uint64_t          len;
 
-    if (size > (ssize_t)(sizeof(intptr_t) * 8)) {
+    if (size > (sizeof(intptr_t) * 8)) {
         abort();
     }
